requirements.txt
torch
torchvision
umap-learn
hdbscan
pillow
numpy
certifi
PyYAML
